Gemini/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
pypdf
groq
langchain-groq
# chromadb
faiss-cpu
langchain_google_genai
langchain-community
pdf2image
pathlib
youtube_transcript_api
crewai
crewai_tools
